first ply compiler
Running main.go will compile test.ply to Go and print the result to
stdout.

The basic process is:
- Parse and type-check test.ply (remember, we modified the type-
  checker so that it knows about 'merge').
- Walk the AST, looking for ocurrences of 'merge'.
  - For each occurrence, lookup the argument types for that
    expression (this is why we need to do type-checking).
  - Replace the occurrence with its 'specialized' name, e.g.
    merge -> mergeintint.
  - Generate a specialized function declaration for the occurrence.
- Add each new declaration to the ast.File.
- Print the ast.File to stdout.

I'm no expert when it comes to the go/ packages, so there's probably
a cleaner/easier way to accomplish this. Regardless, it works, and
it's not to hard to imagine even adding optimizations. For example,
in the following code:

    m1 = merge(m1, m2)

Our specialized 'merge' function could simply copy m2's elements
into m1 (instead of allocating a new map, as it currently does). For
the most part, though, I plan to leave such optimizations to the Go
compiler. In a similar vein, it seems obvious that passing 'nil' in
place of either map argument should be valid; this requires a slight
tweak to the type-checker. But as I stated previously, it remains to
be seen whether 'merge' will make the final cut anyway. For now, it
is simply a proof-of-concept.