invoke Go compiler on output
Also change the codegen a little to merge all the generated code
via ast.MergePackageFiles. I wanted to stick everything in one
giant file, but that screwed up the import statements. Easy to
revisit later.

One sad consequence of this is that we can't output the generated
code to stdout anymore, since it's not a single file. But that's
not a huge deal since you can't pipe to the go command anyway.