implement filter
This is another big step forward: custom methods! These are a bit
trickier because there's no precedent for them in the standard
library; all the builtins are top-level functions, not methods.

Manual testing is working fine for now, but eventually I should
write some automated tests, probably using golden files and
capturing the output of the compiled binaries.

I'll also point out here that the Visit code is pretty ugly. That's
not an accident: I'm waiting until I add more functions and methods
before I start refactoring. Otherwise it's hard to get a clear
picture of what the most appropriate abstraction is.