implement morph
We are closing in on the full picture here. 'morph' is about as
difficult as these functions will get: it's a builtin method with
complicated type-checking. Now that it's working, I can start
cleaning things up and implementing the rest of the builtins.

I implemented morph using a bit of a hack. See, 'filter' was easy
because its type signature contains only one unknown:

    ([]T).filter(func(T) bool) []T

Since we know the value that filter is being called on, we can
produce a full types.Signature for it. This is how most method
lookups work.

However, morph has two unknowns:

    ([]T).morph(func(T) U) []U

This is a problem because at the time when we type-check the
selector ([]T), we haven't evaluated the argument types yet.
Consequently, we can't produce a full types.Signature for this
use of morph. Instead, we have to delay somehow until we have
access to the arguments.

I accomplished this by adding an extra field to types.Signature,
containing the plyId for the method. This ID is then checked in
check.call. If it's non-zero, we call check.plySpecialMethod
instead of proceeding with the usual function call checks. Then
we can do the sort of arbitrary type-checking that we do in
check.ply.

But wait, there's another problem: check.call doesn't have access
to the original selector! Without this, we again can't properly
type-check morph. So we actually need to shove the selector type
into the types.Signature, and extract it later.

Long story short, it's kind of an ugly hack. I'm confident that
there's a cleaner way to accomplish this, but for now I'm just
glad that it works.

Examining the method-related code made me realize that there are
some design decisions left to make, such as how to handle embedded
fields. For example, given the type:

    type foo struct {
        []int
    }

Should we expect it to be legal to call foo.filter? I don't see
why not; but I don't think my current implementation supports this.
Also, what if foo already has a method named filter? I think the
logical thing would be for it to take precedence over the builtin
filter. I switched around the ordering in check.selector in order
to support this behavior.