implement reduce
reduce introduces another piece of trickiness: untyped constants.
For example, consider this reduction:

    xs := []int{1, 2, 3}
    product := func(x, y int) int { return x * y }
    xs.filter(product, 1)

In this expression, '1' is an untyped constant. During type-checking,
we need to convert this constant to the appropriate type. But which
type should we use? The most appropriate is probably the return type
of the reducer function, although the first parameter may also work.
Fortunately there is a convertUntyped method that makes this easy.