implement reassignment optimization
This turned out to be easier than I thought. I played around with
the design a little before deciding that the generator should be
in charge of determining when+how to perform the reassignment
optimization. As a consequence, the generator also needs to be
in charge of rewriting the callsite, so now all the gen functions
need to return a function that can be applied to the top node.
Alternatively, I could have passed that node as an argument to
the gen functions, but I performing all the modifications inside
Visit, as well as only passing the gen functions what they need.

Unfortunately, I found that one of the optimizations I wanted to
apply is not possible. I wanted to optimize merge like this:

    func merge(m1, m2, reassign map[T]U) map[T]U {
        if reassign == m1 {
            // just insert values from m2
        } else if reassign == m2 {
            // just insert values from m1
        } else {
            // insert from both m1 and m2
        }
    }

But it turns out that maps cannot be directly compared. :(
So now I'm considering changing merge to always insert into m1.
Then, if you want to avoid polluting an existing map, you can
pass a make(map[T]U) (or maybe nil?) in as the first argument.