add black-box testing
This is a pretty extreme example of "test the behavior, not the
implementation!" During setup, we compile the ply binary. Then,
for each test case, we write the code to a temp file, call ply
on it, and run the generated binary. Finally, we compare the
output to the expected output for the test case. I like this
approach, but it may make certain things difficult -- such as
comparing the memory usage of optimized vs. unoptimized methods.

Also, this approach made the shortcomings of the ply command quite
painfully obvious. You can even pass a -o flag for the compiled
binary, so I'm forced to cd in and out of temp dirs.