support optional argument to reduce
This wasn't as hard as I expected it to be. I was basing my code
off of the 'make' builtin, which takes a type expression as its
first argument. The standard arg-checking code is only meant to
handle expressions, so 'make' has to do some manual checking.
'reduce', of course, only takes expressions, so it's much easier.
The only similarity is that both 'make' and 'reduce' are variadic.