add reverse
I was surprised to find that reverse is not trivially optimizable.
See the comment in reverseGen for an explanation.

However, I did have an idea for an optimization: in the following
expression, we should be able to avoid any allocations:

    for _, x := range xs.reverse() {
        // use x
    }

Since the reversed slice isn't being stored in a variable anywhere,
we can simply rewrite the loop like so:

    for i := range xs {
        x := xs[len(xs)-i-1]
        // use x
    }

This is a pretty hefty AST rewrite though, so I'll probably put it
off until later.

Also note that when only the index is used, as in:

    for i := range xs.reverse() {
        // use x
    }

We can delete the 'reverse' call entirely, since it doesn't have
any side effects and the length of xs is the same.