add zip
More ambitious would be for zip to accept any number of slices,
provided that the function has an equal number of arguments. For
example, this should be valid:

    sum3 := func(x, y, z int) int { return x + y + z }
    zip(sum3, xs, yz, zs)

The type-checking aspect isn't too hard there; it's the codegen
that's annoying. I would need to repeat some strings n times or
something; it could get ugly.