add contains (for slice and map)
'contains' is the first method implemented on maps. The type-
checking code stayed fairly clean, which is nice. I simply added
a type switch, as well as a new path for checking that the type is
comparable. The codegen is a bit ugly for my tastes, but acceptable.

I had originally planned to implement 'contains' in terms of 'any':

    func (xs []T) contains(e T) bool {
        return xs.any(func(x T) bool {
            return x == e
        })
    }

And in fact, it turned out to be fairly easy to do this: simply re-
run the compiler on the generated code. But when I benchmarked the
result, I found that it was significantly slower than writing out
the full loop. I guess this isn't a huge suprise -- Go does not seem
to optimize function literals very well. At any rate, I removed the
double-compilation code. Despite its convenience, it's best to write
out the implementation of each function in full.