rename reduce -> fold
I think fold is slightly more correc than reduce: the convention
seems to be that reduce does not take an initial value. And honestly
I just prefer the name, probably because I have a soft spot for
Haskell. :P

I'm not sure if it's worthwhile to add a foldr as well. JavaScript
has a "reduceRight" so it's probably common enough to be worth
adding. I'll hold off until there's a pressing need for it, though.