add max/min
These functions started out simple but because very interesting when
I realized that there was an opportunity for compile-time evaluation.
Basically, if the arguments of max/min are both constants, then the
result can be a constant too, since the go/constant package allows
you to compare constants at compile time.

I'm not sure the code is totally correct; there's some weirdness
around converting between untyped constants. For example, you might
expect max(0.1, 1) to evaluate to 1, but instead you get a type
mismatch error: floats can't be compared to ints. So should the
compiler automatically promote 1 to a float? I'm not sure, but for
now I'm going with "no," mostly because Go in general is very
conservative with regard to automatic type conversions.

The constant rewriting is a bit of a hack, as described in the
comment. I'd love to have an easy way of doing it properly. Maybe
there is a function somewhere for returning the parent of a given
ast.Node.