clean up generator templates
This could probably be accomplished with 'text/template' as well,
but I find that syntax to be a little heavy. #T is super clear.

I also replaced safeIdent with randFnName and randTypeName. This has
one clear downside, which is that we always generate new code even
if we could reuse an existing implementation. For example, the code
for a fold that sums ints and a fold that multiplies them is
identical. This isn't a huge deal, though, and dedup seems easy
enough to restore later if it becomes a problem.