implement pipelining
After much toil, I have finally implemented pipelining and all the
(current) tests are passing. This code was initially pretty hideous,
but I amended the commit a few times out of shame. I still consider
it ugly, but at least it's not resume-damaging ugly.

The implementation of pipelining is described in a comment at the top
of pipeline.go. It seems to be working pretty well, but I worry that
the design won't be adequate to support later features. The two most
blatant problems I see now are 1) pipelining only works for slice-to-
slice transformations, and 2) pipelining only works for methods, not
functions. There are also two gross special cases for reverse and
fold1. Nevertheless, I'm proud that the code works. Next, I plan to
separate the codegen from the ply binary, which should make testing
a little easier.

Fixes #2.