add foreach and tee
foreach wasn't worth adding until pipelining was implemented, and
later I realized the value of tee in providing a window into the
contents of a pipeline.
foreach is actually unecessary, since there's no allocation cost
to using tee. So I may combine these later. But for now I like
having them be distinct operations.

I was also curious to see how painful it now is to add new methods.
Currently, to add a method, I have to do the following:

1. Add the method to the methodGenerators map
2. Write the method template and generator
3. Write the transformation template
4. Add the method name and signature to the type-checker
5. Write a test that verifies the method's behavior
6. Update the README and doc.go

This is a bit much. It would be nice if I could at least be spared
the redundancy of defining two separate templates. I have toyed with
the idea of removing the standalone templates and doing all codegen
via the pipeline templates. This has obvious appeal, but it removes
some optimization opportunities. For example, in a standalone morph,
we can preallocate a slice of the exact length necessary. But in a
pipeline, there may be fewer elements in the final slice, so no
preallocation is possible. Perhaps some additional metadata could
be added to enable this, but I worry that the transformation type is
already complex enough.

The best case would be a system where I could just write the template
as plain Go code, with comments demarcating the various sections.
Then some helper function would parse it and split it into pieces.
Each template would live in its own file, so another helper could
scan all the files and generate a table containing the generator
function for each template.

But there is a terrible cost to such a system, and I'm not talking
about the development time required. No, the cost is that it would
make it incredibly easy to add new methods! On the contrary, I want
to make defining new methods as annoying as possible, in order to
discourage the impulse to add frivolous ones. My goal is to add a
set of methods that is as small as possible, but no smaller.