split out codegen and improve ply cmd
ply now supports most common operations: run, build, install, test.
The trickiest part was the flags: I wanted to be able to write:

    ply build -o foo

i.e. use the exact same flags as the go command. But I don't think
there's an easy way of doing this. In order to skip over the flags
properly, we need a way to distinguish what flags are booleans. So
we'd have to import all the flags available to every go subcommand,
*and* somehow keep them in sync with whatever local version of go
the user is running. Not feasible.

We could also half-ass it and make an educated guess. One way would
be to check after each flag whether the rest of the arguments look
like files or packages. But this clearly fails for cases like:

    ply build -o foo.ply

Which *should* produce a binary called foo.go from the current dir,
but would instead interpret -o as a binary flag and try to build only
foo.ply.

So instead, I settled on an ugly solution: having ply take a single
flag, -goflags, which it then passes to the go command. So now to
produce a named binary, you have to do:

    ply -goflags "-o foo" build

Pretty gross, and I don't even want to think about what it looks like
when one of the flags is -ldflags... Still, it works. Maybe there's
a better solution waiting to be discovered someday.

I also questioned whether making ply a compiler binary really made
sense, or if it should be a pure translation tool. The latter has
clear appeal due to convenience, but also some deficiencies (like
the one discussed above). I think it's probably a good idea to add
a pure translation mode, like 'ply compile foo.ply' -- this would
integrate better with some build systems.

I also waffled on the API of the codegen package. It was quite a bit
harder to design than I expected! It will almost certainly need
further refinement, but it's serviceable for now.

Fixes #5