write tests in ply
This style of testing is a lot easier on the eyes! Of course, it
requires a working ply binary, but I think that's acceptable. I'll
eventually augment this with more direct testing of the codegen
package. It would also be nice to write some of these tests with
the testing/quick package.

One thing I didn't anticipate is that when a test fails, you'll
get the line number of the compiled .go file, not the .ply file.
It would be nice if this could be fixed via the //line directive.