add //line directive to codegen
The plan here is to never change the number of lines an expression
occupies unless absolutely necessary. Currently the only exceptions
to this rule are max and min. So we can either add a //line directive
after the max/min call, or perform the rewrite so as to preserve the
total number of lines. While the latter sounds appealing, it winds up
being rather tricky due to multi-line string constants. So a //line
directive is probably the best way to go.

Interestingly, using go/printer with Mode: SourcePos does not produce
sane-looking code. I pretty much have to use it, since it probably
catches edges cases that aren't on my radar), but it badly mangles
the identation. My guess is that it's because I'm modifying the AST
without properly updating all the Pos markers. Fixing that is
probably The Right Thing To Do but right now I'm just too lazy. At
any rate, the mangling only seems to apply to indentation, which
means it doesn't break anything in terms of line numbers.

Fixes #6