switch from ast.Walk to gorewrite.Rewrite
Rewrite is like Walk, but replaces each node with the return
value of the Rewrite function.

I ran into an interesting bug during this change. At first, I changed
the rewriter functions to return entirely new &ast.Node{}s. But this
resulted in a panic when looking up the type of the new expressions.
I hadn't added an entry in s.types for the new object, so the lookup
was returning nil.
Of course, the type of the replacement expression is the same as the
replacement, so I could easily add a new entry after each rewrite.
But it seemed easier to revert to the old behavior of modifying the
supplied ast.Node in-place. Since the object has the same pointer,
the lookup will succeed, and everything continues to work fine.
I do worry that this is not quite explicit enough, though.