print in RawFormat mode
This seems to fix the weird indentation I was observing earlier. I'm
not entirely sure why, though, which makes me uneasy. My best guess
is that the comments inserted by printer.SourcePos screw up the Pos
values in the existing AST. Regardless, I'm pretty sure it only
affects the identation, not the line number, so it's a purely
cosmetic fix.