add sort
sort is interesting because it requires an import. And unfortunately
this isn't super easy to do. Maybe the most "correct" approach would
be for each codegen function to return a list of packages it needed
to import, but it felt too gross to do that just for one method.
Instead I opted for another special case in the specializer. Even
then, it's not as easy as adding the import to the ast.File; when
the files get merged, the imports don't get special treatment, so
you wind up with imports scattered throughout the file, and the Go
parser doesn't like that. Seems pretty stupid to me that the "merge
files" function doesn't handle this for you, but oh well. Instead we
maintain the list of imports separately, dedup them, and add them all
in at the end.

No pipeline yet. I think the only optimization we can do there is to
sort in-place if there's a sort at the end of the chain. Anything
else requires an intermediate slice. More problematically, currently
we have to add some external definitions (the sort.Interface methods)
in order to enable sorting, and the pipeline design doesn't provide a
way to do this.
However, things get much easier once Go 1.8 is released, as it adds a
new function, sort.Slice. Using sort.Slice will simplify the sorting
code significantly, and should allow the pipeline version to be added
with little trouble.