saner named type handling
This largely reverts commit 2ac2a8da in favor of a simpler solution,
wherein the callsite is cast to a named type instead of modifying the
generated code to return the named type. This solves the problem of
how to handle methods on locally-defined types.

This body of changes makes me a little uneasy, mostly because there
are not a ton of tests for named types. Named types aren't super
common though (and local types even less so), so perhaps it's not
something that I should be spending a lot of time on until real bugs
are discovered.

Fixes #12