make a copy when specifying
This prevents a sneaky bug when using the same transformation twice.
It's sneaky because it only affects the params field. That's because
params is a slice, while the other fields are simple strings. So even
when we copy a transformation by value, we're referencing the same
underlying slice data. Making an explicit copy sidesteps this, though
we still need to be careful to construct the new transformation so as
to avoid reusing any memory.