support maps in pipelines
This isn't full support; only keys() and elems() are implemented. But
just those serve pretty well in most cases, as discussed in #9.

I'll follow this up with full support. It's a bit trickier because we
need to keep track of two variables (key/elem) instead of just one
(elem). You'll see.