rename pipeline vars
In preparation for map pipelining, we need distinct variables for
the element and the key. If we eventually add slice methods that take
an index parameter, the index will be stored in #k.

The quirk with #e in keys_map is a little worrying. I hope that this
quirk doesn't impact any other code.