add enum
I knew this would be tricky, but I didn't expect the tricky part to
be fiddling with constant conversions. The resulting code is quite
simple; I imagine that someone more aquainted with the types package
could have written it much faster than I did.

Originally I designed enum to be more flexible: the 1- and 2- arg
forms would also accept negative ranges. But requiring them to be
positive is easier and seems standard (based on Python's range).

One bit of ugliness here is that arguments like 'a' are converted
to runes, not bytes. It might be possible to get around this by
checking if the constant exceeds 255. I assume that in most cases,
people want a []byte, not a []rune.

There are some other fun tricks that we can pull at compile-time if
the arguments are constants. We could pre-generate the whole slice!
That might be overkill. Another trick is to catch bad ranges, like
enum(-1). It's always better to catch things at compile-time.

I'm really itching to integrate "generator" functions like enum into
the pipelining code. They are just crying out for optimization;
enum(10).fold(sum) shouldn't require any allocations. But I worry
that it may be difficult to integrate, as the pipeline code is only
designed for methods. I'd rather not have to rewrite everything,
unless it's going to be a huge improvement.