install each import before type-checking
This (technically) fixes #15, though it's ugly. Compile used to be a
pure function; now it calls the go command a bunch of times. Yuck.
I suppose this could be done in main, but it doesn't feel like main's
responsibility.

I expect that this will also impact compilation times. Fortunately,
'go install' is instant when the package hasn't changed.